Issue 2715: Created AbstractSystemTest class containing basic utilities to start Pravega services in system tests. Deleted repeated code in many initialize() methods thanks to the utilities in AbstractSystemTest.
