+++ test/system/src/test/java/io/pravega/test/system/AbstractFailoverTests.java
@@ -18,17 +18,19 @@ import io.pravega.client.stream.impl.StreamSegments;
 import io.pravega.common.Exceptions;
 import io.pravega.common.concurrent.Futures;
 import io.pravega.common.util.Retry;
-import io.pravega.test.system.framework.Utils;
 import io.pravega.test.system.framework.services.Service;
 import java.net.URI;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ScheduledExecutorService;
-import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
+
 import static org.junit.Assert.assertTrue;
 
+/**
+ * This class provides useful methods for classes that aim at implementing system tests in which service instance
+ * failures are involved during read and write workloads.
+ */
 @Slf4j
 abstract class AbstractFailoverTests extends AbstractReadWriteTest {
 
@@ -43,54 +45,6 @@ abstract class AbstractFailoverTests extends AbstractReadWriteTest {
     URI controllerURIDirect = null;
     Controller controller;
     ScheduledExecutorService controllerExecutorService;
-
-    static URI startZookeeperInstance() {
-        Service zkService = Utils.createZookeeperService();
-        if (!zkService.isRunning()) {
-            zkService.start(true);
-        }
-        List<URI> zkUris = zkService.getServiceDetails();
-        log.debug("Zookeeper service details: {}", zkUris);
-        return zkUris.get(0);
-    }
-
-    static void startBookkeeperInstances(final URI zkUri) {
-        Service bkService = Utils.createBookkeeperService(zkUri);
-        if (!bkService.isRunning()) {
-            bkService.start(true);
-        }
-        List<URI> bkUris = bkService.getServiceDetails();
-        log.debug("Bookkeeper service details: {}", bkUris);
-    }
-
-    static URI startPravegaControllerInstances(final URI zkUri) throws ExecutionException {
-        Service controllerService = Utils.createPravegaControllerService(zkUri);
-        if (!controllerService.isRunning()) {
-            controllerService.start(true);
-        }
-        Futures.getAndHandleExceptions(controllerService.scaleService(3), ExecutionException::new);
-        List<URI> conUris = controllerService.getServiceDetails();
-        log.info("conuris {} {}", conUris.get(0), conUris.get(1));
-        log.debug("Pravega Controller service  details: {}", conUris);
-        // Fetch all the RPC endpoints and construct the client URIs.
-        final List<String> uris = conUris.stream().filter(uri -> Utils.DOCKER_BASED ? uri.getPort() == Utils.DOCKER_CONTROLLER_PORT
-                : uri.getPort() == Utils.MARATHON_CONTROLLER_PORT).map(URI::getAuthority)
-                                         .collect(Collectors.toList());
-
-        URI controllerURI = URI.create("tcp://" + String.join(",", uris));
-        log.info("Controller Service direct URI: {}", controllerURI);
-        return controllerURI;
-    }
-
-    static void startPravegaSegmentStoreInstances(final URI zkUri, final URI controllerURI) throws ExecutionException {
-        Service segService = Utils.createPravegaSegmentStoreService(zkUri, controllerURI);
-        if (!segService.isRunning()) {
-            segService.start(true);
-        }
-        Futures.getAndHandleExceptions(segService.scaleService(3), ExecutionException::new);
-        List<URI> segUris = segService.getServiceDetails();
-        log.debug("Pravega Segmentstore service  details: {}", segUris);
-    }
 
     void performFailoverTest() throws ExecutionException {
         log.info("Test with 3 controller, segment store instances running and without a failover scenario");
